requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
itsdangerous==2.2.0
requests==2.32.3
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
lightgbm>=4.0,<5
python-multipart==0.0.9
tensorflow>=2.15,<3

websockets>=12.0,<14
openpyxl>=3.1,<4

# Optional (only used when C5_MODE=polymarket)
py-clob-client==0.34.6
web3>=6.0,<8